+++ n_lib.h
@@ -81,10 +81,10 @@ const char *i2cNoteTransaction(char *request, char **response);
 bool i2cNoteReset(void);
 const char *serialNoteTransaction(char *request, char **response);
 bool serialNoteReset(void);
-const char *i2cChunkedReceive(uint8_t *buffer, size_t *size, bool delay, size_t timeoutMs, uint32_t *available);
-const char *i2cChunkedTransmit(uint8_t *buffer, size_t size, bool delay);
-const char *serialChunkedReceive(uint8_t *buffer, size_t *size, bool delay, size_t timeoutMs, uint32_t *available);
-const char *serialChunkedTransmit(uint8_t *buffer, size_t size, bool delay);
+const char *i2cChunkedReceive(uint8_t *buffer, uint32_t *size, bool delay, size_t timeoutMs, uint32_t *available);
+const char *i2cChunkedTransmit(uint8_t *buffer, uint32_t size, bool delay);
+const char *serialChunkedReceive(uint8_t *buffer, uint32_t *size, bool delay, size_t timeoutMs, uint32_t *available);
+const char *serialChunkedTransmit(uint8_t *buffer, uint32_t size, bool delay);
 
 // Hooks
 void NoteLockNote(void);
@@ -101,8 +101,8 @@ const char *NoteI2CTransmit(uint16_t DevAddress, uint8_t* pBuffer, uint16_t Size
 const char *NoteI2CReceive(uint16_t DevAddress, uint8_t* pBuffer, uint16_t Size, uint32_t *avail);
 bool NoteHardReset(void);
 const char *NoteJSONTransaction(char *request, char **response);
-const char *NoteChunkedReceive(uint8_t *buffer, size_t *size, bool delay, size_t timeoutMs, uint32_t *available);
-const char *NoteChunkedTransmit(uint8_t *buffer, size_t size, bool delay);
+const char *NoteChunkedReceive(uint8_t *buffer, uint32_t *size, bool delay, size_t timeoutMs, uint32_t *available);
+const char *NoteChunkedTransmit(uint8_t *buffer, uint32_t size, bool delay);
 bool NoteIsDebugOutputActive(void);
 
 // Utilities
