+++ note.h
@@ -318,19 +318,19 @@ void NoteMD5HashToString(unsigned char *hash, char *strbuf, unsigned long buflen
 // High-level helper functions that are both useful and serve to show developers
 // how to call the API
 #define NOTE_C_BINARY_RX_ALL 0
-const char * NoteBinaryDataDecodedLength(size_t *len);
-const char * NoteBinaryDataEncodedLength(size_t *len);
-const char * NoteBinaryDecode(const uint8_t *inBuf, size_t inLen,
-                              uint8_t *outBuf, size_t *outLen);
-const char * NoteBinaryEncode(const uint8_t *inBuf, size_t inLen,
-                              uint8_t *outBuf, size_t *outLen);
+const char * NoteBinaryDataDecodedLength(uint32_t *len);
+const char * NoteBinaryDataEncodedLength(uint32_t *len);
+const char * NoteBinaryDecode(const uint8_t *inBuf, uint32_t inLen,
+                              uint8_t *outBuf, uint32_t *outLen);
+const char * NoteBinaryEncode(const uint8_t *inBuf, uint32_t inLen,
+                              uint8_t *outBuf, uint32_t *outLen);
 uint32_t NoteBinaryMaxDecodedLengthForBufferSize(uint32_t size);
 uint32_t NoteBinaryMaxEncodedLengthForDecodedLength(uint32_t len);
-const char * NoteBinaryReceive(uint8_t *buffer, size_t bufLen,
-                               size_t offset, size_t *dataLen);
+const char * NoteBinaryReceive(uint8_t *buffer, uint32_t bufLen,
+                               uint32_t offset, uint32_t *dataLen);
 const char * NoteBinaryReset(void);
-const char * NoteBinaryTransmit(uint8_t *data, size_t dataLen,
-                                size_t bufLen, size_t offset);
+const char * NoteBinaryTransmit(uint8_t *data, uint32_t dataLen,
+                                uint32_t bufLen, uint32_t offset);
 uint32_t NoteSetSTSecs(uint32_t secs);
 bool NoteTimeValid(void);
 bool NoteTimeValidST(void);
